chore: Move LcncReduction components to the root reduction package
Moves the reduction engine (LcncReduction, ReductionCarrier, LcncCarrierAlg,
LcncKeyAlg, LcncValueAlg, LcncPhaseAlg, LcncSupport, LcncReductions,
ConfixReducers, ForgeReducers, CrmsReducers, ReducerRegistry, LcncPatchCables)
out of the `borg.trikeshed.lcnc.reduction` package to the
`borg.trikeshed.reduction` package. Updates usages in `LcncFanoutElement`,
`IngestCodec`, and `RollupReducer` to reflect this structural move.
