+++ src/commonMain/kotlin/borg/trikeshed/context/lcnc/LcncFanoutElement.kt
@@ -6,10 +6,10 @@ import borg.trikeshed.context.ElementState
 import borg.trikeshed.context.nuid.Capability
 import borg.trikeshed.context.nuid.Nuid
 import borg.trikeshed.context.nuid.NuidFanoutElement
-import borg.trikeshed.lcnc.reduction.LcncCarrierAlg
-import borg.trikeshed.lcnc.reduction.LcncReduction
-import borg.trikeshed.lcnc.reduction.LcncReductions
-import borg.trikeshed.lcnc.reduction.category
+import borg.trikeshed.reduction.LcncCarrierAlg
+import borg.trikeshed.reduction.LcncReduction
+import borg.trikeshed.reduction.LcncReductions
+import borg.trikeshed.reduction.category
 import kotlinx.coroutines.Job
 
 class LcncFanoutElement(
@@ -37,7 +37,7 @@ class LcncFanoutElement(
         val carrier = if (payload != null) {
             typedCarrierAlg.carrier(payload)
         } else {
-            borg.trikeshed.lcnc.reduction.emptySeriesCarrier()
+            borg.trikeshed.reduction.emptySeriesCarrier()
         }
 
         return typedReduction.execute(carrier)
