+++ src/commonMain/kotlin/borg/trikeshed/reduction/LcncCarrierAlg.kt
@@ -1,4 +1,4 @@
-package borg.trikeshed.lcnc.reduction
+package borg.trikeshed.reduction
 
 import borg.trikeshed.cursor.*
 import borg.trikeshed.lib.*
