+++ src/commonTest/kotlin/borg/trikeshed/reduction/LcncReductionCoreTest.kt
@@ -1,4 +1,4 @@
-package borg.trikeshed.lcnc.reduction
+package borg.trikeshed.reduction
 
 import borg.trikeshed.lib.*
 import kotlin.test.*
